Make bib merger accept SierraRecord and convert them into SierraBibRecord
