+++ common/src/main/scala/uk/ac/wellcome/models/transformable/sierra/SierraBibRecord.scala
@@ -15,4 +15,8 @@ object SierraBibRecord {
       data = data,
       modifiedDate = Instant.parse(modifiedDate)
     )
+
+  def apply(sierraRecord: SierraRecord):SierraBibRecord = {
+    SierraBibRecord(id = sierraRecord.id, data = sierraRecord.data, modifiedDate = sierraRecord.modifiedDate)
+  }
 }
