+++ sierra_adapter/sierra_bib_merger/src/main/scala/uk/ac/wellcome/platform/sierra_bib_merger/services/SierraBibMergerWorkerService.scala
@@ -9,7 +9,7 @@ import uk.ac.wellcome.metrics.MetricsSender
 import uk.ac.wellcome.models.aws.SQSMessage
 import uk.ac.wellcome.sqs.{SQSReader, SQSReaderGracefulException, SQSWorker}
 import uk.ac.wellcome.circe._
-import uk.ac.wellcome.models.transformable.sierra.SierraBibRecord
+import uk.ac.wellcome.models.transformable.sierra.{SierraBibRecord, SierraRecord}
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
@@ -23,8 +23,8 @@ class SierraBibMergerWorkerService @Inject()(
     with Logging {
 
   override def processMessage(message: SQSMessage): Future[Unit] =
-    decode[SierraBibRecord](message.body) match {
-      case Right(record) => sierraBibMergerUpdaterService.update(record)
+    decode[SierraRecord](message.body) match {
+      case Right(record) => sierraBibMergerUpdaterService.update(SierraBibRecord(record))
       case Left(e) =>
         Future {
           logger.warn(s"Failed processing $message", e)
