+++ sierra_adapter/sierra_reader/src/test/scala/uk/ac/wellcome/platform/sierra_reader/services/SierraReaderWorkerServiceTest.scala
@@ -14,11 +14,9 @@ import scala.collection.JavaConversions._
 import io.circe.generic.auto._
 import io.circe.syntax._
 import io.circe.parser.decode
-import uk.ac.wellcome.platform.sierra_reader.flow.{
-  SierraRecord,
-  SierraResourceTypes
-}
+import uk.ac.wellcome.platform.sierra_reader.flow.SierraResourceTypes
 import uk.ac.wellcome.circe._
+import uk.ac.wellcome.models.transformable.sierra.SierraRecord
 import uk.ac.wellcome.platform.sierra_reader.modules.WindowManager
 
 import scala.concurrent.duration._
